Correctly generate aliases for enums
